requirements/prod.txt
# Python Version 3.11
# ----------------------------------------------

-r base.txt

psycopg2==2.9.6
gunicorn==20.1.0
django-storages==1.13.2
django-anymail==10.0
whitenoise==6.0.0
django-cloudinary-storage==0.3.0
cloudinary==1.41.0
